asdf

+++ vaet/single_ion_vaet.py
@@ -137,7 +137,7 @@ class SingleIonVAET(PulseSequence):
         rng = np.random.default_rng()
         
         for i in range(m):
-            amp_wf = np.int32([1.]
+            amp_wf = [1.]
             phase_wf = [0.]
             ram_wf = [np.int32(0)] * 1
             self.turns_amplitude_to_ram(amp_wf, amp_wf, ram_wf)
